crypto: Fix CHERI issues
Fix cheri-related warnings from the compiler and from clang-tidy.

- Use uintptr_t for the data members in parser contexts.
- Use the macros to convert pointers to integers where needed.

Co-developed-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>

+++ crypto/asymmetric_keys/pkcs7_parser.c
@@ -26,7 +26,7 @@ struct pkcs7_parse_context {
 	struct pkcs7_signed_info **ppsinfo;
 	struct x509_certificate *certs;		/* Certificate cache */
 	struct x509_certificate **ppcerts;
-	unsigned long	data;			/* Start of data */
+	uintptr_t	data;			/* Start of data */
 	enum OID	last_oid;		/* Last OID encountered */
 	unsigned	x509_index;
 	unsigned	sinfo_index;
@@ -226,7 +226,7 @@ int pkcs7_note_OID(void *context, size_t hdrlen,
 		char buffer[50];
 		sprint_oid(value, vlen, buffer, sizeof(buffer));
 		printk("PKCS7: Unknown OID: [%lu] %s\n",
-		       (unsigned long)value - ctx->data, buffer);
+		       (unsigned long)value - (unsigned long)ctx->data, buffer);
 	}
 	return 0;
 }
@@ -439,7 +439,7 @@ int pkcs7_extract_cert(void *context, size_t hdrlen,
 
 	if (tag != ((ASN1_UNIV << 6) | ASN1_CONS_BIT | ASN1_SEQ)) {
 		pr_debug("Cert began with tag %02x at %lu\n",
-			 tag, (unsigned long)ctx - ctx->data);
+			 tag, (unsigned long)ctx - (unsigned long)ctx->data);
 		return -EBADMSG;
 	}
